requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
mysqlclient==2.2.4
protobuf==3.20.3
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
